Segundo commit

+++ public/controllers/controller.js
@@ -1,10 +1,15 @@
 var myApp = angular.module('myApp',[]);
 myApp.controller('efacturaCtrl',['$scope','$http', function($scope, $http){
-	console.log("Hola mundo desde el Clientes");
+	console.log("Hola mundo desde el controlador");
 
+/*Aqui esta la variable donde tengo la expresion regular y lo agarro con el ng-pattern*/
+$scope.regex = {
+        rfc: /^[a-zA-Z]{3,4}(\d{6})((\D|\d){3})?$/
+      };
 $scope.productos = [];
 $scope.add = function(p)
     {
+      alert("Prodcto agregado correctamente a la factura :) !");
       var temp = {};
       angular.copy(p, temp);
       $scope.productos.push(temp);
@@ -19,10 +24,16 @@ $scope.add = function(p)
 
 $scope.eliminar = function(c)
     {
+      var decision = window.confirm('Estas seguro de eliminar este producto de la factura ?')
+      if(decision == true){
       $scope.productos.splice(c,1);
       subtotalizar();
       iva();
       totalizar();
+      }
+      else{
+        return false;
+      }
     }
 
 $scope.actualizar_factura = function(i)
@@ -68,16 +79,24 @@ refresh();
 $scope.addCliente = function() {
   console.log($scope.cliente);
   $http.post('/efactura', $scope.cliente).success(function(response) {
+    alert("Cliente ingresado correctamente al sistema :) !");
     console.log(response);
     refresh();
   });
 };
 
 $scope.remover = function(id) {
-	console.log("Remover el cliente con el id: " + id);
-	$http.delete('/efactura/' + id).success(function(response) {
-		refresh();
-	});
+  var decision = window.confirm('Estas seguro de eliminar a este cliente ?');
+	if(decision==true){
+    console.log("Remover el cliente con el id: " + id);
+  	$http.delete('/efactura/' + id).success(function(response) {
+  		refresh();
+
+  	});
+  }
+  else{
+    return false;
+  }
 };	
 
 $scope.editar = function(id){
@@ -89,6 +108,7 @@ $scope.editar = function(id){
 $scope.actualizar = function(){
 	console.log($scope.cliente._id);
 	$http.put('/efactura/' + $scope.cliente._id, $scope.cliente).success(function(response){
+  alert("Datos del cliente actualizados correctamente en el sistema :) !");  
 	refresh();
 	$scope.cliente = "";
 	});
